Added fromNow as a plugin
